requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]>=3.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.8.1
reportlab==4.3.1
streamlit==1.44.1
extra-streamlit-components==0.1.81
requests==2.32.3
email-validator==2.2.0
httpx==0.28.1
pytest==8.3.5
